Make TypedIntegers work with absl::Format.
This avoids the need for a bunch of static casts in places when trying
to print them in error messages.

Bug: dawn:2222
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/161688
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>

+++ generator/templates/dawn/native/ValidationUtils.cpp
@@ -42,7 +42,7 @@ namespace {{native_namespace}} {
                         return {};
                 {% endfor %}
                 default:
-                    return DAWN_VALIDATION_ERROR("Value %i is invalid for {{as_cType(type.name)}}.", static_cast<uint32_t>(value));
+                    return DAWN_VALIDATION_ERROR("Value %i is invalid for {{as_cType(type.name)}}.", value);
             }
         }
 
@@ -53,7 +53,7 @@ namespace {{native_namespace}} {
             if ((value & static_cast<{{namespace}}::{{as_cppType(type.name)}}>(~{{type.full_mask}})) == 0) {
                 return {};
             }
-            return DAWN_VALIDATION_ERROR("Value %i is invalid for {{as_cType(type.name)}}.", static_cast<uint32_t>(value));
+            return DAWN_VALIDATION_ERROR("Value %i is invalid for {{as_cType(type.name)}}.", value);
         }
 
     {% endfor %}
