+++ src/dawn/native/CommandBufferStateTracker.cpp
@@ -409,8 +409,7 @@ MaybeError CommandBufferStateTracker::ValidateBufferInRangeForVertexBuffer(uint3
             DAWN_INVALID_IF(vertexBuffer.usedBytesInStride > bufferSize,
                             "Bound vertex buffer size (%u) at slot %u with an arrayStride of 0 "
                             "is smaller than the required size for all attributes (%u)",
-                            bufferSize, static_cast<uint8_t>(usedSlotVertex),
-                            vertexBuffer.usedBytesInStride);
+                            bufferSize, usedSlotVertex, vertexBuffer.usedBytesInStride);
         } else {
             DAWN_ASSERT(strideCount != 0u);
             uint64_t requiredSize = (strideCount - 1u) * arrayStride + vertexBuffer.lastStride;
@@ -425,8 +424,7 @@ MaybeError CommandBufferStateTracker::ValidateBufferInRangeForVertexBuffer(uint3
                 "Vertex range (first: %u, count: %u) requires a larger buffer (%u) than "
                 "the "
                 "bound buffer size (%u) of the vertex buffer at slot %u with stride %u.",
-                firstVertex, vertexCount, requiredSize, bufferSize,
-                static_cast<uint8_t>(usedSlotVertex), arrayStride);
+                firstVertex, vertexCount, requiredSize, bufferSize, usedSlotVertex, arrayStride);
         }
     }
 
@@ -457,8 +455,7 @@ MaybeError CommandBufferStateTracker::ValidateBufferInRangeForInstanceBuffer(
             DAWN_INVALID_IF(vertexBuffer.usedBytesInStride > bufferSize,
                             "Bound vertex buffer size (%u) at slot %u with an arrayStride of 0 "
                             "is smaller than the required size for all attributes (%u)",
-                            bufferSize, static_cast<uint8_t>(usedSlotInstance),
-                            vertexBuffer.usedBytesInStride);
+                            bufferSize, usedSlotInstance, vertexBuffer.usedBytesInStride);
         } else {
             DAWN_ASSERT(strideCount != 0u);
             uint64_t requiredSize = (strideCount - 1u) * arrayStride + vertexBuffer.lastStride;
@@ -473,8 +470,8 @@ MaybeError CommandBufferStateTracker::ValidateBufferInRangeForInstanceBuffer(
                 "Instance range (first: %u, count: %u) requires a larger buffer (%u) than "
                 "the "
                 "bound buffer size (%u) of the vertex buffer at slot %u with stride %u.",
-                firstInstance, instanceCount, requiredSize, bufferSize,
-                static_cast<uint8_t>(usedSlotInstance), arrayStride);
+                firstInstance, instanceCount, requiredSize, bufferSize, usedSlotInstance,
+                arrayStride);
         }
     }
 
@@ -618,8 +615,7 @@ MaybeError CommandBufferStateTracker::CheckMissingAspects(ValidationAspects aspe
         for (BindGroupIndex i : IterateBitSet(mLastPipelineLayout->GetBindGroupLayoutsMask())) {
             DAWN_ASSERT(HasPipeline());
 
-            DAWN_INVALID_IF(mBindgroups[i] == nullptr, "No bind group set at group index %u.",
-                            static_cast<uint32_t>(i));
+            DAWN_INVALID_IF(mBindgroups[i] == nullptr, "No bind group set at group index %u.", i);
 
             BindGroupLayoutBase* requiredBGL = mLastPipelineLayout->GetFrontendBindGroupLayout(i);
             BindGroupLayoutBase* currentBGL = mBindgroups[i]->GetFrontendLayout();
@@ -633,8 +629,7 @@ MaybeError CommandBufferStateTracker::CheckMissingAspects(ValidationAspects aspe
                 "created by the pipeline. Either use the bind group layout returned by calling "
                 "getBindGroupLayout(%u) on the pipeline when creating the bind group, or "
                 "provide an explicit pipeline layout when creating the pipeline.",
-                mLastPipeline, mBindgroups[i], static_cast<uint32_t>(i), currentBGL,
-                static_cast<uint32_t>(i));
+                mLastPipeline, mBindgroups[i], i, currentBGL, i);
 
             DAWN_INVALID_IF(
                 requiredBGL->GetPipelineCompatibilityToken() == PipelineCompatibilityToken(0) &&
@@ -645,15 +640,14 @@ MaybeError CommandBufferStateTracker::CheckMissingAspects(ValidationAspects aspe
                 "compatible. Use an explicit bind group layout when creating bind groups and "
                 "an explicit pipeline layout when creating pipelines to share bind groups "
                 "between pipelines.",
-                mBindgroups[i], static_cast<uint32_t>(i), currentBGL, mLastPipeline);
+                mBindgroups[i], i, currentBGL, mLastPipeline);
 
             DAWN_INVALID_IF(
                 requiredBGL->GetInternalBindGroupLayout() !=
                     currentBGL->GetInternalBindGroupLayout(),
                 "Bind group layout %s of pipeline layout %s does not match layout %s of bind "
                 "group %s set at group index %u.",
-                requiredBGL, mLastPipelineLayout, currentBGL, mBindgroups[i],
-                static_cast<uint32_t>(i));
+                requiredBGL, mLastPipelineLayout, currentBGL, mBindgroups[i], i);
 
             std::optional<uint32_t> packedIndex = FindFirstUndersizedBuffer(
                 mBindgroups[i]->GetUnverifiedBufferSizes(), (*mMinBufferSizes)[i]);
@@ -683,8 +677,7 @@ MaybeError CommandBufferStateTracker::CheckMissingAspects(ValidationAspects aspe
                 return DAWN_VALIDATION_ERROR(
                     "%s bound with size %u at group %u, binding %u is too small. The pipeline (%s) "
                     "requires a buffer binding which is at least %u bytes.%s",
-                    buffer, bufferSize, static_cast<uint32_t>(i),
-                    static_cast<uint32_t>(bindingNumber), mLastPipeline, minBufferSize,
+                    buffer, bufferSize, i, bindingNumber, mLastPipeline, minBufferSize,
                     (bindingInfo.buffer.type == wgpu::BufferBindingType::Uniform
                          ? " This binding is a uniform buffer binding. It is padded to a multiple "
                            "of 16 bytes, and as a result may be larger than the associated data in "
@@ -702,11 +695,9 @@ MaybeError CommandBufferStateTracker::CheckMissingAspects(ValidationAspects aspe
                 "Writable storage buffer binding aliasing found between %s set at bind group index "
                 "%u, binding index %u, and %s set at bind group index %u, binding index %u, with "
                 "overlapping ranges (offset: %u, size: %u) and (offset: %u, size: %u) in %s.",
-                mBindgroups[a.e0.bindGroupIndex], static_cast<uint32_t>(a.e0.bindGroupIndex),
-                static_cast<uint32_t>(a.e0.bindingIndex), mBindgroups[a.e1.bindGroupIndex],
-                static_cast<uint32_t>(a.e1.bindGroupIndex),
-                static_cast<uint32_t>(a.e1.bindingIndex), a.e0.offset, a.e0.size, a.e1.offset,
-                a.e1.size,
+                mBindgroups[a.e0.bindGroupIndex], a.e0.bindGroupIndex, a.e0.bindingIndex,
+                mBindgroups[a.e1.bindGroupIndex], a.e1.bindGroupIndex, a.e1.bindingIndex,
+                a.e0.offset, a.e0.size, a.e1.offset, a.e1.size,
                 mBindgroups[a.e0.bindGroupIndex]
                     ->GetBindingAsBufferBinding(a.e0.bindingIndex)
                     .buffer);
@@ -719,12 +710,10 @@ MaybeError CommandBufferStateTracker::CheckMissingAspects(ValidationAspects aspe
                 "with subresources (base mipmap level: %u, mip level count: %u, base array layer: "
                 "%u, array layer count: %u) and (base mipmap level: %u, mip level count: %u, base "
                 "array layer: %u, array layer count: %u) in %s.",
-                mBindgroups[a.e0.bindGroupIndex], static_cast<uint32_t>(a.e0.bindGroupIndex),
-                static_cast<uint32_t>(a.e0.bindingIndex), mBindgroups[a.e1.bindGroupIndex],
-                static_cast<uint32_t>(a.e1.bindGroupIndex),
-                static_cast<uint32_t>(a.e1.bindingIndex), a.e0.baseMipLevel, a.e0.mipLevelCount,
-                a.e0.baseArrayLayer, a.e0.arrayLayerCount, a.e1.baseMipLevel, a.e1.mipLevelCount,
-                a.e1.baseArrayLayer, a.e1.arrayLayerCount,
+                mBindgroups[a.e0.bindGroupIndex], a.e0.bindGroupIndex, a.e0.bindingIndex,
+                mBindgroups[a.e1.bindGroupIndex], a.e1.bindGroupIndex, a.e1.bindingIndex,
+                a.e0.baseMipLevel, a.e0.mipLevelCount, a.e0.baseArrayLayer, a.e0.arrayLayerCount,
+                a.e1.baseMipLevel, a.e1.mipLevelCount, a.e1.baseArrayLayer, a.e1.arrayLayerCount,
                 mBindgroups[a.e0.bindGroupIndex]
                     ->GetBindingAsTextureView(a.e0.bindingIndex)
                     ->GetTexture());
